Handle deserialization ourselves
Previously we failed to test that the deserialization and related error
handling actually worked as we were mocking RestClient. This made most
of our tests worthless and led to bugs in code that, at first blush, had
very good unit test coverage. Unfortunately RestSharp does not provide
any infrastructure for unit testing code that depends on it.