NamedEntity.Name: create less objects
NamedEntity.Names return a new dictionary, which is a good idea to avoid returning
something that would give the possibility to change the state of the instance.

On the other hand, we know that NamedEntity.Name isn't going to change the names,
there's hence no need to pay the price for this object creation. Price which is
actually quite high since it uses to create such a new Dictionary in a loop.

According to microbench (on .Net 4.0) it saves an order of magnitude.

Also: I use a local variable rather than referencing _names directly in Linq
to ensure we're thread safe