Rename project files to match namespaces
The csproj files and project directories have been rename to match the
assembly names and namespaces.

The fact that the project files and directories do not match the assembly names and
namespaces have been a constant source of minor troubles.

The build script needed numerous small changes for this to work.

The unit tests and benchmark have been upgraded to netcoreapp1.1 so that
they would run without installing packages that are no longer widely
available for Ubuntu. This should not affect end users.